requirements/common.txt
asgiref>=3.4.0
Django==4.0.6
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-jwt
pytz==2022.1
sqlparse==0.4.2
psycopg2==2.9.3
tzdata==2022.1
print-color==0.4.5